pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-ai-service"
version = "1.0.0"
description = "Stateless face detection and recognition inference service"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0,<4.0",
    "gunicorn>=22.0,<23.0",
    "torch>=2.0,<3.0",
    "torchvision>=0.15,<1.0",
    "opencv-python-headless>=4.8,<5.0",
    "numpy>=1.24,<2.0",
    "Pillow>=10.0,<11.0",
    "boto3>=1.34,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
